Refactor restaurant rating

+++ food-delivery-app/src/main/java/com/mentorship/food_delivery_app/restaurant/dto/response/RestaurantRateResponseDto.java
@@ -9,7 +9,7 @@ public record RestaurantRateResponseDto(
 
         UUID id,
         String title,
-        Double rating,
+        Integer rating,
         String comment,
         String customerFullName,
         Instant createdAt
